Mantis 1216 - Added Grade 300 strand to the strand pool

+++ Include/LRFD/StrandPool.h
@@ -77,10 +77,10 @@ public:
    static lrfdStrandPool* GetInstance();
 
    //------------------------------------------------------------------------
-   const matPsStrand* GetStrand(Int32 key,lrfdVersionMgr::Units units);
+   const matPsStrand* GetStrand(Int64 key,lrfdVersionMgr::Units units);
 
    //------------------------------------------------------------------------
-   const matPsStrand* GetStrand(Int32 key);
+   const matPsStrand* GetStrand(Int64 key);
 
    //------------------------------------------------------------------------
    const matPsStrand* GetStrand(matPsStrand::Grade grade,
@@ -91,7 +91,7 @@ public:
    //------------------------------------------------------------------------
    // Returns the lookup key for pStrand.  If pStrand is not a member of
    // the strand pool, returns -1
-   Int32 GetStrandKey(const matPsStrand* pStrand);
+   Int64 GetStrandKey(const matPsStrand* pStrand);
 
    bool CompareStrands(const matPsStrand* pStrandA, const matPsStrand* pStrandB, bool bCompareGrade = true, bool bCompareType = true, bool bCompareCoating = false, bool bCompareSize = false);
 
@@ -125,8 +125,8 @@ protected:
 private:
    // GROUP: DATA MEMBERS
    static lrfdStrandPool* ms_pInstance;
-   static std::map<Int32, std::shared_ptr<matPsStrand> > ms_USStrand;
-   static std::map<Int32, std::shared_ptr<matPsStrand> > ms_SIStrand;
+   static std::map<Int64, std::shared_ptr<matPsStrand> > ms_USStrand;
+   static std::map<Int64, std::shared_ptr<matPsStrand> > ms_SIStrand;
 
    typedef sysSingletonKillerT<lrfdStrandPool> Killer;
    friend Killer;
@@ -202,10 +202,10 @@ public:
    virtual void Next();
 
    //------------------------------------------------------------------------
-   virtual void Move(Int32 pos);
+   virtual void Move(Int64 pos);
 
    //------------------------------------------------------------------------
-   virtual void MoveBy(Int32 dPos);
+   virtual void MoveBy(Int64 dPos);
 
    //------------------------------------------------------------------------
    operator void*() const;
